requirements.txt
streamlit
pillow
torch
torchvision
ultralytics